Serialization functionality for complex graphs
This API can now serialize complex table structures (cyclic references,
keys that are tables, etc) and even restoring the class meta tables
(may make sense to even serialize metatables that are not belonging to
classes).
Throws errors when encountering functions / threads / userdata... so I guess
this'll need to be extended in the future. But I am pretty glad to have it
available now!